feat: Workshop solution for Module 1
This is my initial workshop solution for Module 1 - Rendiner Logic
I. I think I met all of the design requirements except for one
missing piece. I wasn't able to get the contents of the webpage to
extend to the full height of the screen, so there's a gap at the
bottom.